Alterações no nome e na estrututura dos pacotes
