+++ src/br/com/psc/projetoctb/control/view.java
@@ -1,5 +1,5 @@
 package br.com.psc.projetoctb.control;
 
-public class control {
+public class view {
 
 }
