+++ src/br/com/psc/projetoctb/dao/dao.java
@@ -1,0 +1,5 @@
+package br.com.psc.projetoctb.dao;
+
+public class dao {
+
+}
