+++ src/br/com/psc/projetoctb/fachada/bo.java
@@ -1,0 +1,5 @@
+package br.com.psc.projetoctb.fachada;
+
+public class bo {
+
+}
