+++ src/br/com/psc/projetoctb/model/SolicitaAbastecimento.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package br.com.psc.projetoctb.model.vo;
+package br.com.psc.projetoctb.model;
