+++ src/br/com/psc/projetoctb/model/Veiculo.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package br.com.psc.projetoctb.model.vo;
+package br.com.psc.projetoctb.model;
 
 import java.util.List;
