--- src/br/com/psc/projetoctb/model/bo/bo.java
@@ -1,5 +1,0 @@
-package br.com.psc.projetoctb.model.bo;
-
-public class bo {
-
-}
