--- src/br/com/psc/projetoctb/model/dao/dao.java
@@ -1,5 +1,0 @@
-package br.com.psc.projetoctb.model.dao;
-
-public class dao {
-
-}
