+++ src/br/com/psc/projetoctb/view/control.java
@@ -1,5 +1,5 @@
 package br.com.psc.projetoctb.view;
 
-public class view {
+public class control {
 
 }
